+++ crates/fuel-core/src/service/adapters/fuel_gas_price_provider/tests.rs
@@ -8,9 +8,6 @@ mod producer_gas_price_tests;
 #[cfg(test)]
 mod tx_pool_gas_price_tests;
 
-#[cfg(test)]
-mod graph_ql_gas_price_estimate_tests;
-
 fn build_provider<A>(algorithm: A) -> FuelGasPriceProvider<A>
 where
     A: Send + Sync,